Fix Problem 4 solution
Now works correctly for progressions with non-integer common ratio

+++ lab1/src/main/java/lab1/Problem4.java
@@ -49,7 +49,7 @@ public class Problem4 {
             return true;
         }
 
-        int commonRatio = sequence[1] / sequence[0];
+        float commonRatio = (float) sequence[1] / sequence[0];
 
         for (int i = 2; i < sequence.length; i++) {
             if(sequence[i] != sequence[i-1] * commonRatio) {
